+++ src/math/matrix.js
@@ -199,6 +199,21 @@ export class Matrix33 {
   }
 
   /**
+   * Multiply this matrix by a 3D typed array.
+   *
+   * @param {TypedArray[]} sourceArray The input 3D array.
+   * @param {TypedArray[]} outArray The array to write to.
+   */
+  multiplyTypedArray3D(sourceArray, outArray) {
+    for (let i = 0; i < 3; ++i) {
+      outArray[i] = 0;
+      for (let j = 0; j < 3; ++j) {
+        outArray[i] += this.get(i, j) * sourceArray[j];
+      }
+    }
+  }
+
+  /**
    * Multiply this matrix by a 3D vector.
    *
    * @param {Vector3D} vector3D The input 3D vector.
